Include SDL2 on Windows
This was quite a journey.  CMake is reasonably complicated.  I hope that
this will work when I go to build on OSX and Linux; find_package and
pkg-config should work their magic, but if not, I look forward to making
more edits and having to dive into CMake's documentation.

Shout out to @preshing for his guide that made the CMake documentation
readable in no time at all.