Image On Screen
Brute-force rendering an image on the screen with a simple loader and
hard-coded filename.  FindSDL2_image.cmake included in modules for MSVC,
need one for MacOS now.  Including a rough plan just to give myself a
roadmap.

Next is TTF inclusion so that I can print some words to the screen and
some refactoring to create useful abstractions and clean up the code.
It's starting to get a little unsightly.

(Including a simple png asset for demonstration.  Will need to come up
with a way to pass assets between platforms without sticking everything
in source control.)