Beginning Asset Manager
Attempted to move texture loading into a separate asset manager, and
then I found out that the SDL_Renderer is essentially a potential
wrapper around an OpenGL context which is strictly thread-local.  That
means attempts to use it to load "SDL_Texture" is going to be crossing
no-no memory boundaries if I do that in a separate thread.

I haven't figured out a good solution to this yet.  I'll probably still
load images on a separate thread and then notify to the game loading
state on the main thread that it should spin through the conversion of
SDL_Surface to SDL_Texture (which, I'm assuming, loads the images into
video card RAM / OpenGL context).  The thing that sucks about this is
that it's a blocking operation.

Well...

After some searching of the internets, I have confirmed the above
dilemma.  One OpenGL context per thread.  The recommendation is what I
said:  Load images from disk on a background thread, load to VRAM in a
blocking fashion.  Theoretically the latter is fast enough for a user
not to notice or care.

I mean, this is a toy tetris repository, so I shouldn't care.

But I do.  :|

I think what I'll do is have the AssetManager communicate with the
GameLoad state via a processing queue and just make use of the game
loop.  It may look a little ugly at first, but I should be able to make
my loading screen interpolate its draws while the game loop is slowing
down and churning through VRAM loads.