Clean Up Asset Code, Simplify Project Structure
I deleted the 'cmd' folder.  Everything just lives under "game" now,
until I have a reason to have multiple folders or projects.

This is sort of a YAGNI commit.  I deleted a lot of unused junk that
came from ideas that either don't have an impetus yet or don't work in
the architecture.

I'll be continuing work on the asset manager.