FindSDL2_image MacOS CMake Module
BIG NOTE:
I had to change where the headers pointed for SDL2_image to point to SDL.h on
MacOS, since the frameworks are not set up to play nice together.

Now there's a working build and configuration on both Mac and Windows.

I haven't tested things on Linux yet, but they should work as long as SDL2 is
available in the package manager and pkg-config does the right thing.