Working Texture Job Queue
Successfully shuffling things from one thread to another with a brute
mutex.  I'll need to check to see if C++ has the same thing as Go's
RWMutex.  I'm sure that it does.  That would help the speed of the job
checking loop.

I'm not sure about having the job take the renderer as an argument for
texture conversion, but since *I* know that happens on the main thread,
I guess it's okay for now.

Still, things are starting to get hairy again.  Once I get an image on
the screen and the loading state doing its job, I need to do a major
refactor and start looking for where I'm doing dumb memory management
things.