Input Component, Entity Separation
Now I have an input component and a drawing component, both of which are
being updated in the game loop.  Given how I had the asset manager set
up, now I have to think about something:

If I'm asynchronously loading resources, how does that work when the
game is running?  If a texture isn't available yet, what gets displayed?
Should this ever happen?  (I'd prefer that it didn't.)

I guess there's the potential for a wireframe / debug draw mode, which
doesn't need a texture to operate.  I should keep this in mind as well.

How should textures work in the draw components then?  It seems natural
to pass an argument to the constructor of the ones that need a texture.
Maybe that's the right call.  If the asset manager hasn't loaded a
particular texture, then the draw component probably shouldn't be trying
to use it.  In special cases, I suppose I could lazy texture load if it
makes sense for that particular component.