Basic Entity Components
Created simple separate base components; input, transform, and draw.
This should serve as an example of how to work within this system.
There is currently a bug with the motion of the block that should be
fixed in the next commit.  Frame interpolation is tricky business.

There are a few other things fixed in the next commit as well- namely,
assets are now being copied into the cmake build directory to be used by
the program.  This happened because I started using CLion and I needed a
way to make the "Run" button work.