First Attempt at State Management
All of a sudden I'm having to manage things on the heap.  This commit is
a combination of a lot of tidying changes and an attempt to pass update
calls down starting at the state at the top of the stack.

It has me considering a more infrequent update that is used to send
messages to inactive states.  In this case, I think I'll have to change
from a stack to a deque to iterate through the states in the container.

The thought I'm going through with right now is that game state changes
are lazily applied.  Enter and exit are called on update cycles.