Game State Lifecycle, Better Asset Management
I feel like I've figured out a better pattern than I was thinking of
before.  Game state lifecycle now works more like "as you expect," with
flexibility given to the game state itself on what it wants to do when
it's loading, when it's entering, when it's exiting, etc.  The lifecycle
now looks like:

- load() -- called every update while the state says isLoading() is true
- enter() -- one update cycle, called after isLoading() is false
- input() -- called on input regardless of state (can choose no-op)
- update() -- called every update while the state says isQuit() is false
- draw() -- called on draw regardless of state (can choose no-op)
- exit() -- one update cycle, called after isQuit() is true

This lifecycle starts when a state is pushed.  This allows flexibility
because the state being pushed can decide to short-circuit the lifecycle
if necessary.

The asset manager now explicitly has a two-update split.  On the main
thread, a message pump is called in order to do texture loading (to be
respectful of the single-threadedness of the GL context), and the
manager has its own sub-thread that is used for processing all other
kinds of jobs.

Right now, the signal I'm using to determine if a state has everything
it wants loaded (by enqueue-ing) is to simply check that the job queues
are empty.  This presumes that only one state is loading or waiting at a
time.  This probably won't be the case in the future, but for now it
works.  In the future, I may want to have an indexed set of queues or
something.  We'll see.