Simple Loading Screen
It turns out that checking to see if the job queues are empty as an
indicator of whether or not loading was done was not a great decision.
It turns out that though you may have pulled all the jobs out of the
kitty, that doesn't mean that the jobs are done executing.  Whoops.

I think I'm going to have to create an `enqueueBundle()` method on the
asset manager.  I just need an easier (and probably) way than iterating
through and checking the global pool for every texture I've asked for in
order to see if I'm done loading.