Fix Text Load Job Timing Bug
I forgot to wait for the TextLoadJob's texture job to finish before
returning from its `execute()` method.  This caused a crash because the
texture that I expected to exist did not.

Edit:  I also should have fixed the interface.  Now there are two
methods on the asset manager: `executeTextureJob` and
`executeTextureJobAsync`

Both submit texture jobs to be processed as part of the game loop, but
the synchronous one does its own polling.