Fix Shutdown Bug
The game was not shutting down properly due to waiting for texture jobs
to get pumped that were never going to again.  Waiting on the thread to
finish in the destructor was a bad idea, since the destructor gets
called at the end of the main and things can be destroyed in *any order*
which means that the mutexes that you thought you had, you don't, and
they're going to crash your program because they just evaporated.

So I now have an explicit shutdown.  I may want to look at this
lifecycle under a microscope in the future, because jobs that are in
progress should probably interrupt themselves more than just expecting
the program to clean everything up.  What if I want to save state before
exiting?