Wrap Renderer in Display Manager
Attempting to cache window dimensions in the display manager and wrap
convenience functions around the `SDL_Renderer`.

Also fixed the "snapping" in the draw component by multiplying by
`remainder_frames - 1`, which accounts for the KEYUP frame.  This is not the
best solution (that I'm aware of), but I'm not sure yet how to account for the
KEYUP frame.  Might try that next.