Modify Block State From UI
I believe this should be a good pattern to establish for the modification of
state from the UI and the passing of state down to each component.  I chose not
to use structs because that would require generics in order to destruct them in
a lifecycle, and I decided it wasn't worth it.  It gets messy.  I tried.